Initial code forked from https://github.com/tomitribe/jkta
This is a temporary codebase with the goal of getting a completely
running TomEE 8 under the jakarta namespace as soon as possible and
without any overhead of abstractions or dealing with third-party
libraries or third-party communities.

This is intended as a stop-gap and this code base will not live
long-term.  Any improvements will be migrated to either the Eclipse
Transformer or Jkta, hopefully making this codebase not necessary.